Move modification of SphinxDocString to conf.py
